Initial work to include layers on client
