+++ src/sinks/datadog/logs/log_api.rs
@@ -1,4 +1,6 @@
 use crate::sinks::datadog::logs::config::Encoding;
+use crate::sinks::datadog::logs::log_api::builder::LogApiBuilder;
+use crate::sinks::datadog::logs::log_api::errors::FlushError;
 use crate::sinks::util::buffer::GZIP_FAST;
 use crate::sinks::util::encoding::{EncodingConfigWithDefault, EncodingConfiguration};
 use crate::sinks::util::Compression;
@@ -10,152 +12,21 @@ use futures::Future;
 use futures::StreamExt;
 use http::{Request, Uri};
 use hyper::Body;
-use snafu::Snafu;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
-use std::hash::{BuildHasherDefault, Hasher};
+use std::hash::BuildHasherDefault;
 use std::io::Write;
 use tokio::time::{self, Duration};
 use tower::Service;
 use twox_hash::XxHash64;
-use vector_core::config::{log_schema, LogSchema};
 use vector_core::event::EventStatus;
 use vector_core::event::{Event, EventFinalizers, LogEvent, Value};
 use vector_core::sink::StreamSink;
 use vector_core::ByteSizeOf;
 
-#[derive(Debug, Snafu)]
-pub enum BuildError {
-    #[snafu(display("The builder is missing the URI to use for Datadog's logs API"))]
-    MissingUri,
-    #[snafu(display("The builder is missing an HTTP client to use Datadog's logs API"))]
-    MissingHttpClient,
-}
-
-#[derive(Debug, Snafu)]
-pub enum LogApiError {}
-
-#[derive(Debug)]
-pub struct LogApiBuilder<Client>
-where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
-    Client::Future: Send,
-    Client::Response: Send,
-    Client::Error: Send,
-{
-    encoding: EncodingConfigWithDefault<Encoding>,
-    http_client: Option<Client>,
-    datadog_uri: Option<Uri>,
-    default_api_key: Option<Box<str>>,
-    compression: Option<Compression>,
-    timeout: Option<Duration>,
-    bytes_stored_limit: u64,
-    log_schema_message_key: Option<&'static str>,
-    log_schema_timestamp_key: Option<&'static str>,
-    log_schema_host_key: Option<&'static str>,
-}
-
-impl<Client> Default for LogApiBuilder<Client>
-where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
-    Client::Future: Send,
-    Client::Response: Send,
-    Client::Error: Send,
-{
-    fn default() -> Self {
-        Self {
-            encoding: Default::default(),
-            http_client: None,
-            datadog_uri: None,
-            default_api_key: None,
-            compression: None,
-            timeout: None,
-            bytes_stored_limit: u64::max_value(),
-            log_schema_message_key: None,
-            log_schema_timestamp_key: None,
-            log_schema_host_key: None,
-        }
-    }
-}
-
-impl<Client> LogApiBuilder<Client>
-where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
-    Client::Future: Send,
-    Client::Response: Send,
-    Client::Error: Send,
-{
-    pub fn log_schema(mut self, log_schema: &'static LogSchema) -> Self {
-        self.log_schema_message_key = Some(log_schema.message_key());
-        self.log_schema_timestamp_key = Some(log_schema.timestamp_key());
-        self.log_schema_host_key = Some(log_schema.host_key());
-        self
-    }
-
-    pub fn encoding(mut self, encoding: EncodingConfigWithDefault<Encoding>) -> Self {
-        self.encoding = encoding;
-        self
-    }
-
-    pub fn default_api_key(mut self, api_key: Box<str>) -> Self {
-        self.default_api_key = Some(api_key);
-        self
-    }
-
-    pub fn bytes_stored_limit(mut self, limit: u64) -> Self {
-        self.bytes_stored_limit = limit;
-        self
-    }
-
-    // TODO enable and set from config
-    // pub fn batch_timeout(mut self, timeout: Duration) -> Self {
-    //     self.timeout = Some(timeout);
-    //     self
-    // }
-
-    pub fn http_client(mut self, client: Client) -> Self {
-        self.http_client = Some(client);
-        self
-    }
-
-    pub fn datadog_uri(mut self, uri: Uri) -> Self {
-        self.datadog_uri = Some(uri);
-        self
-    }
-
-    pub fn compression(mut self, compression: Compression) -> Self {
-        self.compression = Some(compression);
-        self
-    }
-
-    pub fn build(self) -> Result<LogApi<Client>, BuildError> {
-        let mut key_slab = HashMap::default();
-        let default_api_key = self.default_api_key.unwrap();
-        let default_api_key_id = hash(&default_api_key);
-        key_slab.insert(default_api_key_id, default_api_key);
-
-        let log_api = LogApi {
-            default_api_key: default_api_key_id,
-            key_slab,
-            event_batches: HashMap::default(),
-            bytes_stored_limit: self.bytes_stored_limit as usize,
-            bytes_stored: 0,
-            compression: self.compression.unwrap_or_default(),
-            datadog_uri: self.datadog_uri.ok_or(BuildError::MissingUri)?,
-            encoding: self.encoding,
-            http_client: self.http_client.ok_or(BuildError::MissingHttpClient)?,
-            log_schema_host_key: self.log_schema_host_key.unwrap_or(log_schema().host_key()),
-            log_schema_message_key: self
-                .log_schema_message_key
-                .unwrap_or(log_schema().message_key()),
-            log_schema_timestamp_key: self
-                .log_schema_timestamp_key
-                .unwrap_or(log_schema().timestamp_key()),
-            timeout: self.timeout.unwrap_or(Duration::from_secs(60)),
-        };
-        Ok(log_api)
-    }
-}
+mod builder;
+mod common;
+mod errors;
 
 const MAX_PAYLOAD_ARRAY: usize = 1_000;
 
@@ -166,10 +37,71 @@ struct FlushMetrics {
     success: bool,
 }
 
+#[inline]
+fn dissect_batch(batch: Vec<Event>) -> (Vec<BTreeMap<String, Value>>, Vec<EventFinalizers>) {
+    let mut members: Vec<BTreeMap<String, Value>> = Vec::with_capacity(batch.len());
+    let mut finalizers: Vec<EventFinalizers> = Vec::with_capacity(batch.len());
+    for event in batch.into_iter() {
+        let (fields, mut metadata) = event.into_log().into_parts();
+        members.push(fields);
+        finalizers.push(metadata.take_finalizers());
+    }
+    (members, finalizers)
+}
+
+fn build_request(
+    members: Vec<BTreeMap<String, Value>>,
+    api_key: &str,
+    datadog_uri: Uri,
+    compression: &Compression,
+    flush_metrics: &mut FlushMetrics,
+) -> Result<Request<Body>, FlushError> {
+    let total_members = members.len();
+    assert!(total_members <= 1_000);
+    let body: Vec<u8> = serde_json::to_vec(&members).expect("failed to encode to json");
+
+    let request = Request::post(datadog_uri)
+        .header("Content-Type", "application/json")
+        .header("DD-API-KEY", api_key);
+    let serialized_payload_len = body.len();
+    metrics::histogram!("encoded_payload_size_bytes", serialized_payload_len as f64);
+    let (request, encoded_body) = match compression {
+        Compression::None => (request, body),
+        Compression::Gzip(level) => {
+            let level = level.unwrap_or(GZIP_FAST);
+            let mut encoder = GzEncoder::new(
+                Vec::with_capacity(serialized_payload_len),
+                flate2::Compression::new(level as u32),
+            );
+
+            encoder.write_all(&body)?;
+            (
+                request.header("Content-Encoding", "gzip"),
+                encoder.finish().expect("failed to encode"),
+            )
+        }
+    };
+    flush_metrics.processed_bytes_total = serialized_payload_len as u64;
+    flush_metrics.processed_events_total = total_members as u64;
+    request
+        .header("Content-Length", encoded_body.len())
+        .body(Body::from(encoded_body))
+        .map_err(Into::into)
+}
+
+// NOTE likely implementation for #8491
+fn rename_key(from_key: &'static str, to_key: &'static str, log: &mut LogEvent) {
+    if from_key != to_key {
+        if let Some(val) = log.remove(from_key) {
+            log.insert_flat(to_key, val);
+        }
+    }
+}
+
 #[derive(Debug)]
 pub struct LogApi<Client>
 where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
+    Client: Service<Request<Body>> + Send + Unpin,
     Client::Future: Send,
     Client::Response: Send,
     Client::Error: Send,
@@ -234,63 +166,9 @@ where
     compression: Compression,
 }
 
-#[inline]
-fn dissect_batch(batch: Vec<Event>) -> (Vec<BTreeMap<String, Value>>, Vec<EventFinalizers>) {
-    let mut members: Vec<BTreeMap<String, Value>> = Vec::with_capacity(batch.len());
-    let mut finalizers: Vec<EventFinalizers> = Vec::with_capacity(batch.len());
-    for event in batch.into_iter() {
-        let (fields, mut metadata) = event.into_log().into_parts();
-        members.push(fields);
-        finalizers.push(metadata.take_finalizers());
-    }
-    (members, finalizers)
-}
-
-fn build_request(
-    members: Vec<BTreeMap<String, Value>>,
-    api_key: &str,
-    datadog_uri: Uri,
-    compression: &Compression,
-    flush_metrics: &mut FlushMetrics,
-) -> Request<Body> {
-    let total_members = members.len();
-    assert!(total_members <= 1_000);
-    let body: Vec<u8> = serde_json::to_vec(&members).expect("failed to encode to json");
-
-    let request = Request::post(datadog_uri)
-        .header("Content-Type", "application/json")
-        .header("DD-API-KEY", api_key);
-    let serialized_payload_len = body.len();
-    metrics::histogram!("encoded_payload_size_bytes", serialized_payload_len as f64);
-    metrics::gauge!("encoded_payload_size_members", total_members as f64);
-    let (request, encoded_body) = match compression {
-        Compression::None => (request, body),
-        Compression::Gzip(level) => {
-            let level = level.unwrap_or(GZIP_FAST);
-            let mut encoder = GzEncoder::new(
-                Vec::with_capacity(serialized_payload_len),
-                flate2::Compression::new(level as u32),
-            );
-
-            encoder.write_all(&body).expect("failed to write body");
-            (
-                request.header("Content-Encoding", "gzip"),
-                encoder.finish().expect("failed to encode"),
-            )
-        }
-    };
-    flush_metrics.processed_bytes_total = serialized_payload_len as u64;
-    flush_metrics.processed_events_total = total_members as u64;
-    request
-        .header("Content-Length", encoded_body.len())
-        .body(Body::from(encoded_body))
-        // .map_err(Into::into)
-        .expect("failed to make request")
-}
-
 impl<Client> LogApi<Client>
 where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
+    Client: Service<Request<Body>> + Send + Unpin,
     Client::Future: Send,
     Client::Response: Send,
     Client::Error: Send,
@@ -307,7 +185,7 @@ where
     fn register_key_id(&mut self, event: &Event) -> u64 {
         if let Some(api_key) = event.metadata().datadog_api_key() {
             let key = api_key.as_ref();
-            let key_hash = hash(key);
+            let key_hash = common::hash(key);
             // TODO it'd be nice to avoid passing through String
             self.key_slab
                 .entry(key_hash)
@@ -367,14 +245,10 @@ where
     fn flush_to_api<'a>(
         self: &'a mut Self,
         api_key_id: u64,
-    ) -> Result<impl Future<Output = ()> + Send, ()> {
-        let batch = self
-            .event_batches
-            .remove(&api_key_id)
-            .expect("impossible situation");
+        batch: Vec<Event>,
+    ) -> Result<impl Future<Output = ()> + Send, FlushError> {
         assert!(batch.len() <= MAX_PAYLOAD_ARRAY);
 
-        metrics::counter!("flush", 1);
         let api_key = self
             .key_slab
             .get(&api_key_id)
@@ -389,7 +263,7 @@ where
             &self.compression,
             &mut flush_metrics,
         );
-        let fut = self.http_client.call(request).map(move |result| {
+        let fut = self.http_client.call(request?).map(move |result| {
             let status: EventStatus = match result {
                 Ok(_) => {
                     metrics::counter!("flush_success", 1);
@@ -414,26 +288,10 @@ where
     }
 }
 
-#[inline]
-fn hash(input: &str) -> u64 {
-    let mut hasher = XxHash64::default();
-    hasher.write(input.as_bytes());
-    hasher.finish()
-}
-
-// NOTE likely implementation for #8491
-fn rename_key(from_key: &'static str, to_key: &'static str, log: &mut LogEvent) {
-    if from_key != to_key {
-        if let Some(val) = log.remove(from_key) {
-            log.insert_flat(to_key, val);
-        }
-    }
-}
-
 #[async_trait]
 impl<Client> StreamSink for LogApi<Client>
 where
-    Client: Service<Request<Body>> + Send + Clone + Unpin,
+    Client: Service<Request<Body>> + Send + Unpin,
     Client::Future: Send,
     Client::Response: Send,
     Client::Error: Send,
@@ -460,13 +318,7 @@ where
 
         let mut interval = time::interval(self.timeout);
         let mut flushes = FuturesUnordered::new();
-        metrics::gauge!("bytes_stored_limit", self.bytes_stored_limit as f64);
         loop {
-            metrics::gauge!("outstanding_requests", flushes.len() as f64);
-            metrics::gauge!("bytes_stored", self.bytes_stored as f64);
-            // TODO the current ticking method does not take into account
-            // flushing. That is, no matter what, we will always flush every
-            // interval even if we just finished flushing.
             tokio::select! {
                 biased;
                 Some(()) = flushes.next() => {
@@ -476,18 +328,15 @@ where
                     let key_id = self.register_key_id(&event);
                     let event_size = event.size_of();
                     if !self.has_space(key_id) || self.bytes_stored_limit < self.bytes_stored + event_size {
-                        // We've gone over the bytes limit and must flush our
-                        // batches. As a future optimization, since we haven't
-                        // passed the timeout yet, we might consider only
-                        // flushing a single large batch. As of self writing the
-                        // batch structure doesn't allow sorting by size so we
-                        // just ship the whole thing.
-                        //
+                        let batch = self
+                            .event_batches
+                            .remove(&key_id)
+                            .expect("impossible situation");
                         // TODO currently we do serialization inline to this
                         // call, which is time consuming relative to processing
                         // incoming events. If we could serialize in parallel to
                         // doing IO we'd end up ahead most likely.
-                        let flush = self.flush_to_api(key_id)?;
+                        let flush = self.flush_to_api(key_id, batch).expect("unsure how to handle error");
                         flushes.push(flush);
                         self.bytes_stored = 0; // TODO logic is now wrong since we flush single-batch
                     }
@@ -495,14 +344,23 @@ where
                     self.bytes_stored += event_size;
                 }
                 _ = interval.tick() => {
+                    // The current ticking method does not take into account
+                    // flushing. That is, no matter what, we will always flush
+                    // every interval even if we just finished flushing. We
+                    // could avoid this if we were able to keep track of when a
+                    // batch was last flushed _or_ if we pull batching
+                    // responsibility (+ timeouts) into a stream manipulator
+                    // prior to this sink.
                     let keys: Vec<u64> = self.event_batches.keys().into_iter().map(|x| *x).collect();
                     for key_id in keys.into_iter() {
-                        let flush = self.flush_to_api(key_id)?;
+                        let batch = self
+                            .event_batches
+                            .remove(&key_id)
+                            .expect("impossible situation");
+                        let flush = self.flush_to_api(key_id, batch).expect("unsure how to handle error");
                         flushes.push(flush);
                     }
                     self.bytes_stored = 0;
-                    // TODO
-                    // flushes.extend(self.flush_to_api().await?)
                 },
             }
         }
